+++ server/core/src/indexer/utils.ts
@@ -31,20 +31,22 @@ import core, {
   FullTextSearchContext,
   generateId,
   Hierarchy,
-  IndexedReader,
   IndexStageState,
   isFullTextAttribute,
   isIndexedAttribute,
   Obj,
   Ref,
+  RefTo,
   Space,
   Storage,
   TxFactory,
-  ClassSearchConfigProps
+  ClassSearchConfigProps,
+  docKey
 } from '@hcengineering/core'
 import { deepEqual } from 'fast-equals'
 import plugin from '../plugin'
-import { FullTextPipeline } from './types'
+import { FullTextPipeline} from './types'
+import type { IndexedDoc } from '../types'
 /**
  * @public
  */
@@ -304,6 +306,14 @@ export function collectPropagateClasses (pipeline: FullTextPipeline, objectClass
   return Array.from(propagate.values())
 }
 
+/**
+ * @public
+ */
+export interface IndexedReader {
+  get: (attribute: string) => any
+  getDoc: (attribute: string) => IndexedReader | undefined
+}
+
 export function readAndMapProps (reader: IndexedReader, props: ClassSearchConfigProps[]): { [key: string]: string } {
   const res: { [key: string]: string } = {}
   for (const prop of props) {
@@ -323,4 +333,32 @@ export function readAndMapProps (reader: IndexedReader, props: ClassSearchConfig
 
 export function fillTemplate (tmpl: string, props: { [key: string]: string }): string {
   return tmpl.replace(/{(.*?)}/g, (_, key: string) => props[key])
+}
+
+/**
+ * @public
+ */
+export function createIndexedReader (
+  _class: Ref<Class<Doc>>,
+  hierarchy: Hierarchy,
+  doc: IndexedDoc,
+  refAttribute?: string
+): IndexedReader {
+  return {
+    get: (attr: string) => {
+      const realAttr = hierarchy.findAttribute(_class, attr)
+      if (realAttr !== undefined) {
+        return doc[docKey(attr, { refAttribute, _class: realAttr.attributeOf })]
+      }
+      return undefined
+    },
+    getDoc: (attr: string) => {
+      const realAttr = hierarchy.findAttribute(_class, attr)
+      if (realAttr !== undefined) {
+        const refAtrr = realAttr.type as RefTo<Doc>
+        return createIndexedReader(refAtrr.to, hierarchy, doc, docKey(attr, { _class }))
+      }
+      return undefined
+    }
+  }
 }
